tts_krutrim/requirements.txt
gradio>=3.50.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0